Config file modifications
Recent commits increased the number of config parameters and the parameters are no longer loaded to param server. The config files are modified to make it easier to work with these changes, by modifying parameters which needed another func parameter for populating to a larger list, are now dicts with func and value fields. config_utils.py is modified to address this.

README updated