New topo maps for the simulation.
Topo map for testing row traversal in simulation adjusted so that all in row edges have action `row_traversal`.
UV version of above topo map made for the tall robot.

Param of row detector `min_num_poles` set to zero for use at Riseholme.
Scenario for tall robot uv: the param `use_omni` is set to true as it is unlikely this robot will be able to make uturns at the end of the rows.
Small change in script `add_edges_between_nodes` that lets the user set the edge action.