Multi-Ellipse Row Detector.
Information in each ellipse is processed independently.
The config for an ellipse is specified as a (dictionary) item in a list in `row_detector_ellipse_cfg.yaml`
Currently this config is set up for the fruit transportation scenario and replicates the behaviour of the previous row detector config.
There is also a config for the UV scenario (`row_detector_ellipse_uv-cfg.yaml`) with 4 ellipses (3 for row detection, 1 for obstacle detection).
Parameters that are currently common to all ellipses are specified in a separate config file `row_detector_common_cfg.yaml`.

Some of the recent changes to the RASberry stack broke `setup_corner_hokuyos.launch` which has now been fixed.