Improvements to rasberry optimise.
The overall change is that parameter sets are now specified as lists of dictionaries
and stored in yaml files (json files are no longer used in this package).

Each dictionary item in the list represents a parameter.
The namespace of the parameter is a key in this dictionary which is used to
create reconfigure clients.

Most of the changes in this commit were made to accommodate the above.